Tishchuk-fix: moved Auth.js and Registration.js to the new folder - "helpers" in src. Renamed Registration.js to RegistrationHandler.js. Renamed RegistrationForm.js to Registration.js. Added consts instead of lets. Changed user1 to initialUser in RegistrationHandler.js. Added Olivias' recent version of jsx in Registration.js
