+++ src/helpers/Auth.js
